gnn-traffic-colab/src/models/lstm_baseline.py
To put your project into Google Colab, you can follow these steps:

1. **Upload Your Files**: You can upload your project files directly to Google Colab. Use the file upload feature in Colab to upload your `.py` files, data files, and any other necessary resources.

2. **Mount Google Drive**: If your project files are stored in Google Drive, you can mount your Google Drive in Colab. This allows you to access files directly from your Drive.

   Use the following code to mount your Google Drive:

   from google.colab import drive
   drive.mount('/content/drive')

3. **Install Required Libraries**: Make sure to install any libraries that your project depends on. You can use pip commands in a Colab cell to install them.

   For example:
   !pip install pandas numpy torch torch-geometric

4. **Run Your Code**: You can copy and paste your code into a Colab cell or create a new Python file in the Colab environment. You can then run your code as you would in any Python environment.

5. **Save Your Work**: If you make changes to your code or data, you can save them back to your Google Drive or download them to your local machine.

By following these steps, you should be able to run your project in Google Colab without any issues.