gnn-traffic-colab/requirements_colab.txt
# Google Colab Requirements for GNN Traffic Project
# Optimized for Colab environment with compatible versions

# Core ML libraries (Colab compatible versions)
torch>=1.12.0
torch-geometric>=2.1.0
torch-audio>=0.12.0
torch-vision>=0.13.0

# Essential data processing
numpy>=1.21.0
pandas>=1.4.0
scipy>=1.8.0
scikit-learn>=1.1.0

# Graph processing
networkx>=2.8.0

# Geospatial data (lightweight for Colab)
geopandas>=0.11.0
shapely>=1.8.0
pyproj>=3.3.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.8.0
folium>=0.12.0

# Configuration and utilities
pyyaml>=6.0
tqdm>=4.64.0
requests>=2.28.0

# Colab specific
google-colab
ipywidgets>=7.7.0

# Optional for advanced features
# cartopy>=0.20.0  # Uncomment if needed for advanced mapping